Started material editor

+++ Engine/Utility/Time.h
@@ -7,7 +7,7 @@ struct Time
 	static float delta_time;
 	static float time;
 
-	static float frame_duration();
+	static uint32_t frame_duration(); // in ms
 
 	struct Chrono
 	{
